+++ convex-peer/src/main/java/convex/net/impl/nio/Connection.java
@@ -204,7 +204,7 @@ public class Connection extends AConnection {
 		while (!clientChannel.finishConnect()) {
 			long now = Utils.getCurrentTimestamp();
 			long elapsed=now-start;
-			if (elapsed > Config.DEFAULT_CLIENT_TIMEOUT)
+			if (elapsed > Config.DEFAULT_INTERNAL_TIMEOUT)
 				throw new TimeoutException("Couldn't connect after "+elapsed+"ms");
 			try {
 				Thread.sleep(10+elapsed/3);